requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Audio Processing
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
pydub==0.25.1

# Speech Recognition & Language Detection
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# Speaker Recognition
speechbrain==0.5.16
pyannote.audio==3.1.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
